+++ app/feature/feature-claim-chat/src/androidMain/kotlin/com/hedvig/feature/claim/chat/ClaimChatNavGraph.kt
@@ -90,7 +90,7 @@ fun NavGraphBuilder.claimChatGraph(
   }
   navdestination<ClaimOutcomeDeflectDestination>(ClaimOutcomeDeflectDestination) {
     ClaimOutcomeDeflectDestination(
-      deflect = deflect,
+      deflect = deflect.deflectData,
       imageLoader = imageLoader,
       navigateUp = navController::navigateUp,
       openUrl = openUrl,
